sizeof working
